// src/app/shared/model/UploadType.ts
export enum UploadType {
  Unknown = 0,
  ExtenuatingCircumstanceProof = 1,
  RegistrationFormSupportingDocuments = 2,
  PreRegistrationFormSupportingDocuments = 3,
  PharmacyTechnicianSupportingDocuments = 4,
  ProgressReportSupportingDocuments = 5,
  AssessmentRegistrationDocuments = 6,
  InpendentPrescriberSupportDocuments = 7,
  ReturnToRegisterSupportingDocuments = 8,
  ProfessionalStandingSupportingDocuments = 9
}